Clean things up
